Add field->lookup() which chains down to datatype->lookup().  This allows us to verify enumerations, etc.
Also make regaddr_t be 64 bits - mmio could have 64 bit values.

Add some FIXMEs to consider later, when they might matter.

+++ trunk/pp.h
@@ -86,6 +86,6 @@ typedef enum pp_bitwidth {
 /*
  * pp_regaddr - the address of a register.
  */
-typedef uint32_t pp_regaddr;
+typedef uint64_t pp_regaddr;
 
 #endif // PP_PP_H__
